finished splitting SDK logic between web3 and widget - testing needed
